User Registration 90% Completed

+++ admin/includes/view_all_posts.php
@@ -153,7 +153,7 @@ if(isset($_POST['select_submit']) && isset($_POST['checkBoxArray']))
   echo "<td>{$post_comments}</td>";
   echo "<td>{$post_date}</td>";
   echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-  echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+  echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?'); \"href='posts.php?delete={$post_id}'>Delete</a></td>";
   echo "<td><a href='../post.php?p_id={$post_id}'>View</a></td>";
   echo "</tr>";
